# Lab2/openapi.yaml
openapi: 3.0.3
info:
  title: Computer Store
  version: 1.0.0
...
paths:
...
  /users/login:
    get:
      tags:
        - users
      summary: Logs users into the system
      description: ''
      operationId: loginusers
      parameters:
        - name: login
          in: query
          schema:
            type: string
            description: login
          required: false
        - name: password
          in: query
          required: false
          schema:
            type: string
            description: password
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Invalid login or password
        '500':
          description: Status internal server error
...
  /users/register:
    post:
      tags:
        - users
      summary: Register
      description: Add a new users
      operationId: registerusers
      requestBody:
        description: Create a new users
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users'
          application/xml:
            schema:
              $ref: '#/components/schemas/users'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
  /users:
    get:
      tags:
        - users
      summary: Get all userss
      description: Get users all userss
      operationId: GetAlluserss
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/users'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/users'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
  /users?username=string:
    get:
      tags:
        - users
      summary: Get users by usersname
      description: Get users by usersname
      operationId: GetusersByName
      parameters:
        - name: usersname
          in: query
          description: The name that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/users'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/users'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
  /users/{id}:
    get:
      tags:
        - users
      summary: Get users by id
      description: Get users by id
      operationId: GetusersById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/users'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/users'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - users
      summary: Updates a users with form data
      description: ''
      operationId: updateusers
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    patch:
      tags:
        - users
      summary: Partially update a specific users by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Fields to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users'
      responses:
        '200':
          description: products partially updated successfully
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: products not found
    delete:
      tags:
        - users
      summary: Deletes a users
      description: delete a users
      operationId: deleteusers
      parameters:
        - name: id
          in: path
          description: users id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
...
  /carts:
    post:
      tags:
        - carts
      summary: Add a new carts
      description: Add a new carts
      operationId: addcarts
      requestBody:
        description: Create a new carts
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/carts'
          application/xml:
            schema:
              $ref: '#/components/schemas/carts'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /carts? users_id = int:
    get:
      tags:
        - carts
      summary: Get carts by users_id
      description: Get carts by users_id
      operationId: GetcartsByIdusers
      parameters:
        - name: users_id
          in: query
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
  /carts/{id}:
    get:
      tags:
        - carts
      summary: Get carts by id
      description: Get carts by id
      operationId: GetcartsById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - carts
      summary: Updates carts
      description: ''
      operationId: updatecarts
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/carts'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - carts
      summary: Deletes a carts
      description: delete a carts
      operationId: deletecarts
      parameters:
        - name: id
          in: path
          description: users id to carts
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /itemcarts:
    post:
      tags:
        - itemcarts
      summary: Add a new item carts
      description: Add a new itemcarts
      operationId: addItemcarts
      requestBody:
        description: Create a new item carts
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemcarts'
          application/xml:
            schema:
              $ref: '#/components/schemas/itemcarts'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /itemcarts/{id}:
    get:
      tags:
        - itemcarts
      summary: Get itemcarts by id
      description: Get itemcarts by id
      operationId: GetItemcartsById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/itemcarts'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/itemcarts'
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - itemcarts
      summary: Updates itemcarts
      description: ''
      operationId: updateItemcarts
      parameters:
        - name: id
          in: path
          description: ID of Itemcarts that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemcarts'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - itemcarts
      summary: Deletes a itemcarts
      description: delete a itemcarts
      operationId: deleteItemcarts
      parameters:
        - name: id
          in: path
          description: users id to itemcarts
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
components:
  schemas:
...
    carts:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        users_id:
          type: integer
          format: int64
          example: 1
        data_created:
          type: string
          format: date-time
          example: '2024-09-20'
      xml:
        name: carts
    itemcarts:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        carts_id:
          type: integer
          format: int64
          example: 1
        products_id:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          example: 1
      xml:
        name: itemcarts
...
    users:
      type: object
      properties:
        usersname:
          type: string
          example: theusers
        phone:
          type: string
          example: '+79267197512'
        login:
          type: string
          example: James
        password:
          type: string
          example: '12345'
        email:
          type: string
          example: 12345@gmail.com
        address:
          type: string
          example: Moscow
        role:
          type: string
          enum:
            - admin
            - client
            - guest
          example: client
      xml:
        name: users